:truck: Rename validSudoku.js to validifiers.js
